collectePython/requirements.txt
pymongo
requests
schedule
python-dotenv
